+++ src/surround.rs
@@ -7,46 +7,69 @@
 // your option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-//! Surround speaker configurations and types.
-//!
-//! # Channel Locations
-//!  0. Front Left
-//!  1. Front Right
-//!  2. Back Left
-//!  3. Back Right - Surround 4.0
-//!  4. Front Center
-//!  5. LFE (low frequency effects) - Surround 5.1
-//!  6. Side Left
-//!  7. Side Right - Surround 7.1
+//! Surround Sound 5.1 speaker configuration and types.
 
 use crate::{
-    chan::{Ch16, Ch32, Ch64, Ch8},
-    sample::{Sample4, Sample6, Sample8},
+    chan::{Ch16, Ch32, Ch64, Ch8, Channel},
+    sample::Sample,
 };
 
-/// 4.0 Surround [8-bit PCM](../chan/struct.Ch8.html) format.
-pub type Surround4x8 = Sample4<Ch8>;
-/// 4.0 Surround [16-bit PCM](../chan/struct.Ch16.html) format.
-pub type Surround4x16 = Sample4<Ch16>;
-/// 4.0 Surround [32-bit Floating Point](../chan/struct.Ch32.html) format.
-pub type Surround4x32 = Sample4<Ch32>;
-/// 4.0 Surround [64-bit Floating Point](../chan/struct.Ch64.html) format.
-pub type Surround4x64 = Sample4<Ch64>;
+/// Surround Sound 5.1 sample format (front left channel, rear left channel,
+/// rear right, front right channel, center, lfe).
+#[derive(Default, PartialEq, Copy, Clone, Debug)]
+#[repr(transparent)]
+pub struct Surround<C: Channel> {
+    channels: [C; 6],
+}
+
+impl<C: Channel> Surround<C> {
+    /// Create a one-channel Sample.
+    pub fn new<H>(one: H, two: H, three: H, four: H, five: H, six: H) -> Self
+    where
+        C: From<H>,
+    {
+        let channels = [
+            C::from(one),
+            C::from(two),
+            C::from(three),
+            C::from(four),
+            C::from(five),
+            C::from(six),
+        ];
+        Self { channels }
+    }
+}
+
+impl<C: Channel> Sample for Surround<C> {
+    const CONFIG: &'static [[f64; 2]] = &[
+        [1.0 / 12.0, 0.25],         // Front Left (Centered at 1/6)
+        [0.25, 0.5],                // Rear Left (Centered at 1/3)
+        [0.5, 0.75],                // Rear Right (Centered at 2/3)
+        [0.75, 11.0 / 12.0],        // Front Right (Centered at 5/6)
+        [11.0 / 12.0, 13.0 / 12.0], // Front Center
+        [f64::NAN, f64::NAN],       // LFE
+    ];
+
+    type Chan = C;
+
+    fn channels(&self) -> &[Self::Chan] {
+        &self.channels
+    }
+
+    fn channels_mut(&mut self) -> &mut [Self::Chan] {
+        &mut self.channels
+    }
+
+    fn from_channels(ch: &[Self::Chan]) -> Self {
+        Self::new::<C>(ch[0], ch[1], ch[2], ch[3], ch[4], ch[5])
+    }
+}
 
 /// 5.1 Surround [8-bit PCM](../chan/struct.Ch8.html) format.
-pub type Surround5x8 = Sample6<Ch8>;
+pub type Surround8 = Surround<Ch8>;
 /// 5.1 Surround [16-bit PCM](../chan/struct.Ch16.html) format.
-pub type Surround5x16 = Sample6<Ch16>;
+pub type Surround16 = Surround<Ch16>;
 /// 5.1 Surround [32-bit Floating Point](../chan/struct.Ch32.html) format.
-pub type Surround5x32 = Sample6<Ch32>;
+pub type Surround32 = Surround<Ch32>;
 /// 5.1 Surround [64-bit Floating Point](../chan/struct.Ch64.html) format.
-pub type Surround5x64 = Sample6<Ch64>;
-
-/// 7.1 Surround [8-bit PCM](../chan/struct.Ch8.html) format.
-pub type Surround7x8 = Sample8<Ch8>;
-/// 7.1 Surround [16-bit PCM](../chan/struct.Ch16.html) format.
-pub type Surround7x16 = Sample8<Ch16>;
-/// 7.1 Surround [32-bit Floating Point](../chan/struct.Ch32.html) format.
-pub type Surround7x32 = Sample8<Ch32>;
-/// 7.1 Surround [64-bit Floating Point](../chan/struct.Ch64.html) format.
-pub type Surround7x64 = Sample8<Ch64>;
+pub type Surround64 = Surround<Ch64>;